+++ packages/swap/src/adapters/intl/number-locale.mocks.ts
@@ -1,0 +1,10 @@
+export const mockNumberLocale = {
+  decimalSeparator: '.',
+  fractionGroupSeparator: ' ',
+  fractionGroupSize: 0,
+  groupSeparator: ',',
+  groupSize: 3,
+  prefix: '',
+  secondaryGroupSize: 0,
+  suffix: '',
+} as const
