+++ packages/types/src/helpers/types.ts
@@ -1,12 +1,1 @@
 export type Nullable<T> = T | null | undefined
-
-export type NumberLocale = {
-  prefix: string
-  decimalSeparator: string
-  groupSeparator: string
-  groupSize: number
-  secondaryGroupSize: number
-  fractionGroupSize: number
-  fractionGroupSeparator: string
-  suffix: string
-}
