+++ packages/types/src/intl/numbers.ts
@@ -1,0 +1,10 @@
+export type NumberLocale = {
+  prefix: string
+  decimalSeparator: string
+  groupSeparator: string
+  groupSize: number
+  secondaryGroupSize: number
+  fractionGroupSize: number
+  fractionGroupSeparator: string
+  suffix: string
+}
